+++ NEWS.md
@@ -1,3 +1,7 @@
+* Added descriptive error messages when unknown `Flow.aggregation.point`, 
+  `Flow`, or `Product` are encountered in `arrange_iea_fu_allocation_template()`.
+
+
 # IEATools 0.1.60 (2021-09-06) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5476175.svg)](https://doi.org/10.5281/zenodo.5476175)
 
 * No longer discriminating by sign when summing EIOU after